Transform fixes, physics simulation time fix, start of Editor
- Now '@transform =' has semantics of setting the transform, whereas @transform() applies a relative transform.
- Fixed some randomness in the physics engine based on setting transforms
- Transform listeners are no longer notified if they are the originator of the transform. This is very importatn
  for physics where sometimes we were getting updated from Bullet, and immediately re-updating Bullet.
- Physics has its own time delta calculation, which cleaned up the simulation
- Messages are processed before physics, after physics, and after Entity update. I believe this has removed
  a 1-frame delay between pressing a key and seeing movement.  There is still at least 1 frame between but
  based on the async-nature of gaen, I haven't conceived a fix for this.
- Time is now f64 until passed into Entity update. Thus, Bullet gets a higher resolution time. Still need
  f32 Entity update since Compose doesn't support f64 as a type.
- Editor class added, can listen to kepresses, not much else yet
- Editor is created based on DevBuild/command arguments to gaen
- Keypress support added to input processing
- Now, instead of just being able to poll for complex states, you can also get events when keypress/keyrepeats happen
- Fixed issue where multiple input modes couldn't be specified simultaneously. Now, input.conf can have sections and
  it's possible to switch between modes.

+++ src/core/logging.cpp
@@ -98,7 +98,7 @@ void Logger::log(LogSeverity severity, const char * message)
     // will become un-necessary.
     LogMessage &lm = tLogMessage;
 
-    lm.header.time = now();
+    lm.header.time = (f32)now();
     lm.header.sev = severity;
 
     size_t msgLen = max(kMaxLogMessageSize-1, strlen(message) + 1);
